# src/main/resources/openapi/license_decider.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/licenses/add/":{"post":{"summary":"Add License","operationId":"add_license_licenses_add__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/License"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/":{"get":{"summary":"Get All New Licenses","description":"Returns all available software licenses as License-Objects","operationId":"get_all_new_licenses_licenses__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All New Licenses Licenses  Get","type":"array","items":{"$ref":"#/components/schemas/License"}}}}}}}},"/licenses/{license_id}":{"get":{"summary":"Get License","operationId":"get_license_licenses__license_id__get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update License","operationId":"update_license_licenses__license_id__put","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/License"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete License","operationId":"delete_license_licenses__license_id__delete","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/conditions":{"get":{"summary":"Get License Conditions","operationId":"get_license_conditions_licenses__license_id__conditions_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/limitations":{"get":{"summary":"Get License Limitations","operationId":"get_license_limitations_licenses__license_id__limitations_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/permissions":{"get":{"summary":"Get License Permissions","operationId":"get_license_permissions_licenses__license_id__permissions_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/check/":{"post":{"summary":"Check License","operationId":"check_license_licenses_check__post","requestBody":{"content":{"application/json":{"schema":{"title":"Usedlicenses","type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Check License Licenses Check  Post","type":"array","items":{"$ref":"#/components/schemas/License"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Condition":{"title":"Condition","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"License":{"title":"License","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"url":{"title":"Url","maxLength":65536,"minLength":1,"type":"string","format":"uri"},"type":{"$ref":"#/components/schemas/LicenseType"},"conditions":{"title":"Conditions","type":"array","items":{"$ref":"#/components/schemas/Condition"},"default":[]},"permissions":{"title":"Permissions","type":"array","items":{"$ref":"#/components/schemas/Permission"},"default":[]},"limitations":{"title":"Limitations","type":"array","items":{"$ref":"#/components/schemas/Limitation"},"default":[]}}},"LicenseType":{"title":"LicenseType","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"Limitation":{"title":"Limitation","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"Permission":{"title":"Permission","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}